Added in shape classes for rendering
