+++ lib/cli.js
@@ -1,35 +1,11 @@
 const inquirer = require('inquirer');
 const MaxLengthInputPrompt = require('inquirer-maxlength-input-prompt');
-const generateSVG = require('./generateSVG.js')
+const generateSVG = require('./generateSVG.js');
+const { Circle, Triangle, Square } = require('../shapes/shapes.js');
+const questions = require('./questions.js');
 const fs = require('fs');
 
-inquirer.registerPrompt('maxlength-input', MaxLengthInputPrompt)
-
-// An array of questions for the user to answer
-const questions = [
-  {
-    type: 'maxlength-input',
-    name: 'text',
-    message: 'Please enter 3 characters for your logo',
-    maxLength: 3
-  },
-  {
-    type: 'input',
-    name: 'textColor',
-    message: 'What color would you like the characters to be?',
-  },
-  {
-    type: 'list',
-    name: 'shapeSelect',
-    message: 'Pick a shape',
-    choices: ['Circle', 'Triangle', 'Square']
-  },
-  {
-    type: 'input',
-    name: 'shapeColor',
-    message: 'which color would you like for your shape?'
-  }
-]
+inquirer.registerPrompt('maxlength-input', MaxLengthInputPrompt);
 
 class CLI {
   constructor() {
@@ -54,20 +30,21 @@ class CLI {
         // Switch case that will match the users shape selection
         switch (shapeSelect) {
 
-          // Based on the users selection, it will create a new class of the desired shape and console log their selection while saving it the 'userShape' variable
+          // Based on the users selection, it will create and render a new class of the desired shape and console log their selection while saving it the 'userShape' variable
            case "Circle":
             console.log(log);
-            userShape = `<circle cx="150" cy="100" r="95" height="100%" width="100%" fill="${shapeColor}" /><text x="150" y="115" font-size="60" text-anchor="middle" fill="${textColor}">${text}</text>`;
+            userShape = new Circle(text, textColor, shapeColor).render();
+            console.log(userShape);
             break;
 
            case "Triangle":
             console.log(log);
-            userShape = `<polygon height="100%" width="100%" points="20,200 275,200 150,10" fill="${shapeColor}" /><text x="150" y="160" font-size="60" text-anchor="middle" fill="${textColor}">${text}</text>`
+            userShape = new Triangle(text, textColor, shapeColor).render();
             break;
 
            case "Square":
             console.log(log);
-            userShape = `<rect x="90" y="40" width="150" height="150" fill="${shapeColor}" /><text x="165" y="130" font-size="60" text-anchor="middle" fill="${textColor}">${text}</text>`
+            userShape = new Square(text, textColor, shapeColor).render();
             break;
         }
